Switch to CMake, Siganim now builds for Windows
This commit switches build systems, now relying on CMake instead of
Eclipse CDT's managed builder.

Using MXE, we can build Siganim with either static or dynamic linking; a
static executable is about 73MB (closer to 50 when stripped), whereas a
dynamic one is about 3MB along with a 90MB DLL package.

Two switches are available for this project:

  - STATIC_LINKING, that instructs the build system to use static
versions of ICU (which names begin with an "s") - defaults to OFF,
  - GUI_ENABLE, that builds Siganim with its Qt GUI - defaults to ON.
Disabling this option removes the need for Qt.

+++ src/sign/SignCell.h
@@ -17,6 +17,7 @@
 #ifndef SRC_SIGNCELL_H_
 #define SRC_SIGNCELL_H_
 
+#include <inttypes.h>
 #include <iostream>
 #include <exception>
 #include "SignTree.h"
@@ -59,9 +60,9 @@ public:
     public:
         NoSuchChildException(const SignCell* sender,
                 const SignCell* child) : sender(sender), child(child) {
-            this->message = "Cell at " + std::to_string((unsigned long)sender) +
+            this->message = "Cell at " + std::to_string((uintptr_t)sender) +
                     " has no such child " +
-                    std::to_string((unsigned long)child);
+                    std::to_string((uintptr_t)child);
         }
 
         const char* what() const throw () {
@@ -78,7 +79,7 @@ public:
         std::string message;
     public:
         OrphanNodeException(const SignCell* sender) : sender(sender) {
-            this->message = "Cell at " + std::to_string((unsigned long)sender) +
+            this->message = "Cell at " + std::to_string((uintptr_t)sender) +
                     " is orphan ";
         }
